Implement unify mode
Unify mode uses -P to suppress line directives in the preprocessed
output. The dependency file generated by '-MD' is additionally used to
track included files to be hashed for direct mode.

Additionally, -fnormalize-whitespace is used if supported by the
compiler (Clang 13 and later).

The following configuration options are added:

 * unify_mode to enable/disabled unify mode (disabled by default)

 * preprocessor, preprocessor_check, preprocessor_type to use a
   different preprocessor to determine whether a recompilation is
   necessary

 * incbin sloppiness to skip scanning for incbin directives

 * unify_with_debug sloppiness to allow matching unified hashes
   when generating debug info

 * unify_with_output sloppiness to allow matching unified hashes
   even f the compiler emitted diagnostic messages

The patch also fixes some issues when running tests, especially when
running on Windows.